wsd: test: merge UnitWopiHttpRedirectLoop with Redirect
